Revert "Merge pull request #1509 from boardsesh/native-tab-bar"
This reverts commit 396f14258c38305d12a7f7227f0342e3beef90ca, reversing
changes made to 5ee861ad3f1d23fd8e4748fdec0065348a21467c.
