+++ mobile/ios/App/App/SceneDelegate.swift
@@ -4,31 +4,30 @@ import Capacitor
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
     var window: UIWindow?
+    /// Universal link URL received during cold start, before the Capacitor bridge is ready.
+    var pendingUniversalLinkURL: URL?
 
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let windowScene = (scene as? UIWindowScene) else { return }
 
         // Save any universal link that triggered this cold start.
-        // The webviews aren't ready yet, so we store it and let
-        // MultiWebViewController pick it up after viewDidLoad.
-        var pendingURL: URL?
+        // The bridge isn't ready yet, so we store it and let
+        // BoardseshViewController pick it up after viewDidLoad.
         if let userActivity = connectionOptions.userActivities.first(where: {
             $0.activityType == NSUserActivityTypeBrowsingWeb
         }), let url = userActivity.webpageURL {
-            pendingURL = url
+            pendingUniversalLinkURL = url
         }
 
         let window = UIWindow(windowScene: windowScene)
         let storyboard = UIStoryboard(name: "Main", bundle: nil)
-        let rootVC = storyboard.instantiateInitialViewController()
-        if let multiWebVC = rootVC as? MultiWebViewController, let url = pendingURL {
-            multiWebVC.pendingUniversalLinkURL = url
-        }
-        window.rootViewController = rootVC
+        window.rootViewController = storyboard.instantiateInitialViewController()
         self.window = window
         window.makeKeyAndVisible()
 
         // Clean up any orphaned Live Activities from a previous session/crash.
+        // Use Task.detached to avoid inheriting MainActor context and blocking
+        // the main run loop with Activity<> enumeration continuations.
         if #available(iOS 16.1, *) {
             Task.detached(priority: .utility) {
                 await LiveActivityManager.shared.endStaleActivities()
@@ -55,17 +54,17 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
             restorationHandler: { _ in }
         )
 
-        // Navigate the correct tab's webview for universal links (warm start).
+        // Navigate the WebView directly for universal links (warm start).
+        // Without @capacitor/app plugin, the proxy alone does not trigger navigation.
         if userActivity.activityType == NSUserActivityTypeBrowsingWeb,
            let url = userActivity.webpageURL,
-           let multiVC = window?.rootViewController as? MultiWebViewController {
-            let tab = MultiWebViewController.tabForPath(url.path)
-            let targetTab = (tab == "create") ? "climbs" : tab
-            multiVC.navigateToTab(targetTab, url: url.absoluteString)
+           let vc = window?.rootViewController as? BoardseshViewController {
+            vc.webView?.load(URLRequest(url: url))
         }
     }
 
     func sceneDidDisconnect(_ scene: UIScene) {
+        // End Live Activity when the scene is discarded to avoid stale state
         SessionWebSocketManager.shared.disconnect()
         if #available(iOS 16.1, *) {
             Task.detached(priority: .utility) {
@@ -75,6 +74,10 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     }
     func sceneDidBecomeActive(_ scene: UIScene) {}
     func sceneWillResignActive(_ scene: UIScene) {}
-    func sceneWillEnterForeground(_ scene: UIScene) {}
+    func sceneWillEnterForeground(_ scene: UIScene) {
+        // Live Activities auto-dismiss after their 3-minute stale date.
+        // Cold-start cleanup in scene(_:willConnectTo:) handles the rest.
+        // No work needed here — avoids ActivityKit IPC on every foreground.
+    }
     func sceneDidEnterBackground(_ scene: UIScene) {}
 }
