+++ packages/web/app/components/queue-control/queue-control-bar.tsx
@@ -73,8 +73,6 @@ import { shareWithFallback } from '@/app/lib/share-utils';
 import { getPreference, setPreference } from '@/app/lib/user-preferences-db';
 import styles from './queue-control-bar.module.css';
 import { PLAY_DRAWER_EVENT as PLAY_DRAWER_EVENT_INTERNAL, dispatchOpenPlayDrawer } from './play-drawer-event';
-import { isNativeApp } from '@/app/lib/ble/capacitor-utils';
-import { addNativeOverlay, removeNativeOverlay } from '@/app/lib/native-tab-bar/native-tab-bar-plugin';
 
 export type ActiveDrawer = 'none' | 'play' | 'queue' | 'tick';
 
@@ -175,13 +173,6 @@ const QueueControlBar: React.FC<QueueControlBarProps> = ({ boardDetails, angle }
     window.addEventListener(PLAY_DRAWER_EVENT_INTERNAL, handler);
     return () => window.removeEventListener(PLAY_DRAWER_EVENT_INTERNAL, handler);
   }, []);
-
-  // Notify native tab bar to hide while any drawer covers the screen (ref-counted with BottomTabBar).
-  useEffect(() => {
-    if (!isNativeApp() || activeDrawer === 'none') return;
-    addNativeOverlay();
-    return () => removeNativeOverlay();
-  }, [activeDrawer]);
 
   // Scroll to current climb when drawer finishes opening
   const handleDrawerOpenChange = useCallback((open: boolean) => {
